test: repair sidebar build status CI gates
