+++ src/agents/embedded-agent-runner/runs.test.ts
@@ -11,8 +11,6 @@ import {
   getDiagnosticSessionState,
   resetDiagnosticSessionStateForTest,
 } from "../../logging/diagnostic-session-state.js";
-import { createUserTurnTranscriptRecorder } from "../../sessions/user-turn-transcript.js";
-import { createTestUserTurnTranscriptTarget } from "../../sessions/user-turn-transcript.test-support.js";
 import {
   abortAndDrainEmbeddedAgentRun,
   abortEmbeddedAgentRun,
@@ -21,48 +19,10 @@ import {
   isEmbeddedAgentRunAbortableForRunId,
   isEmbeddedAgentRunAbortableForCompaction,
   isEmbeddedAgentRunHandleActive,
-  formatEmbeddedAgentQueueFailureSummary,
-  queueEmbeddedAgentMessageWithOutcome,
-  queueEmbeddedAgentMessageWithOutcomeAsync,
   retainEmbeddedAgentRunAbortabilityForRunId,
   setActiveEmbeddedRun,
 } from "./runs.js";
-import { testing } from "./runs.test-support.js";
-
-type RunHandle = Parameters<typeof setActiveEmbeddedRun>[1];
-
-function createRunHandle(
-  overrides: {
-    abort?: () => void;
-    isAbortable?: boolean;
-    isCompacting?: boolean;
-    isStreaming?: boolean;
-    isStopped?: () => boolean;
-    messageInjection?: RunHandle["messageInjection"];
-    runId?: string;
-    queueMessage?: RunHandle["queueMessage"];
-    supportsQueueMessageImages?: boolean;
-    supportsTranscriptCommitWait?: boolean;
-  } = {},
-): RunHandle {
-  // Minimal handle fixture with overrideable lifecycle probes for registry
-  // behavior; individual tests supply queue/abort behavior when needed.
-  const abort = overrides.abort ?? (() => {});
-  return {
-    runId: overrides.runId,
-    queueMessage: overrides.queueMessage ?? (async () => {}),
-    ...(overrides.messageInjection ? { messageInjection: overrides.messageInjection } : {}),
-    isStreaming: () => overrides.isStreaming ?? true,
-    ...(overrides.isStopped ? { isStopped: overrides.isStopped } : {}),
-    ...(overrides.isAbortable !== undefined
-      ? { isAbortable: () => overrides.isAbortable !== false }
-      : {}),
-    isCompacting: () => overrides.isCompacting ?? false,
-    supportsQueueMessageImages: overrides.supportsQueueMessageImages,
-    supportsTranscriptCommitWait: overrides.supportsTranscriptCommitWait,
-    abort,
-  };
-}
+import { createEmbeddedRunHandle, testing } from "./runs.test-support.js";
 
 describe("embedded-agent runner run registry", () => {
   afterEach(() => {
@@ -81,10 +41,10 @@ describe("embedded-agent runner run registry", () => {
 
     setActiveEmbeddedRun(
       "session-compacting",
-      createRunHandle({ isCompacting: true, abort: abortCompacting }),
+      createEmbeddedRunHandle({ isCompacting: true, abort: abortCompacting }),
     );
 
-    setActiveEmbeddedRun("session-normal", createRunHandle({ abort: abortNormal }));
+    setActiveEmbeddedRun("session-normal", createEmbeddedRunHandle({ abort: abortNormal }));
 
     const aborted = abortEmbeddedAgentRun(undefined, { mode: "compacting" });
     expect(aborted).toBe(true);
@@ -110,9 +70,12 @@ describe("embedded-agent runner run registry", () => {
     const abortA = vi.fn();
     const abortB = vi.fn();
 
-    setActiveEmbeddedRun("session-a", createRunHandle({ isCompacting: true, abort: abortA }));
+    setActiveEmbeddedRun(
+      "session-a",
+      createEmbeddedRunHandle({ isCompacting: true, abort: abortA }),
+    );
 
-    setActiveEmbeddedRun("session-b", createRunHandle({ abort: abortB }));
+    setActiveEmbeddedRun("session-b", createEmbeddedRunHandle({ abort: abortB }));
 
     const aborted = abortEmbeddedAgentRun(undefined, { mode: "all" });
     expect(aborted).toBe(true);
@@ -122,7 +85,7 @@ describe("embedded-agent runner run registry", () => {
 
   it("keeps finalizing runs active while rejecting abort requests", () => {
     const abort = vi.fn();
-    const handle = createRunHandle({ abort, isAbortable: false });
+    const handle = createEmbeddedRunHandle({ abort, isAbortable: false });
     const operation = createReplyOperation({
       sessionKey: "agent:main:finalizing",
       sessionId: "session-finalizing",
@@ -155,7 +118,7 @@ describe("embedded-agent runner run registry", () => {
 
   it("keeps frozen run ownership through forced in-process restart", () => {
     const abort = vi.fn();
-    const handle = createRunHandle({ abort, isAbortable: false });
+    const handle = createEmbeddedRunHandle({ abort, isAbortable: false });
     const operation = createReplyOperation({
       sessionKey: "agent:main:restart-finalizing",
       sessionId: "session-restart-finalizing",
@@ -185,7 +148,7 @@ describe("embedded-agent runner run registry", () => {
   });
 
   it("binds abortability to the owning run id", () => {
-    const finalizing = createRunHandle({
+    const finalizing = createEmbeddedRunHandle({
       abort: vi.fn(),
       isAbortable: false,
       runId: "run-finalizing",
@@ -203,7 +166,7 @@ describe("embedded-agent runner run registry", () => {
     clearActiveEmbeddedRun("session-shared", finalizing);
     expect(isEmbeddedAgentRunAbortableForRunId("run-finalizing")).toBe(false);
 
-    const queued = createRunHandle({ runId: "run-queued" });
+    const queued = createEmbeddedRunHandle({ runId: "run-queued" });
     setActiveEmbeddedRun("session-shared", queued);
 
     expect(isEmbeddedAgentRunAbortableForRunId("run-finalizing")).toBe(false);
@@ -215,7 +178,7 @@ describe("embedded-agent runner run registry", () => {
 
   it("passes restart ownership to every aborted run", () => {
     const abort = vi.fn();
-    setActiveEmbeddedRun("session-restart", createRunHandle({ abort }));
+    setActiveEmbeddedRun("session-restart", createEmbeddedRunHandle({ abort }));
 
     expect(abortEmbeddedAgentRun(undefined, { mode: "all", reason: "restart" })).toBe(true);
     expect(abort).toHaveBeenCalledWith("restart");
@@ -257,7 +220,7 @@ describe("embedded-agent runner run registry", () => {
       sessionId: "session-reply-stuck-live",
       resetTriggered: false,
     });
-    const handle = createRunHandle({
+    const handle = createEmbeddedRunHandle({
       abort: () => {
         operation.abortByUser();
       },
@@ -284,7 +247,7 @@ describe("embedded-agent runner run registry", () => {
 
   it("claims shared restart ownership before invoking an attached handle", () => {
     const abort = vi.fn();
-    const handle = createRunHandle({ abort });
+    const handle = createEmbeddedRunHandle({ abort });
     const operation = createReplyOperation({
       sessionKey: "agent:main:restart-owned",
       sessionId: "session-restart-owned",
@@ -309,7 +272,7 @@ describe("embedded-agent runner run registry", () => {
     "does not bypass frozen shared ownership through %s handle aborts",
     (mode) => {
       const abort = vi.fn();
-      const handle = createRunHandle({ abort, isCompacting: true });
+      const handle = createEmbeddedRunHandle({ abort, isCompacting: true });
       const sessionId = `session-restart-frozen-${mode}`;
       const operation = createReplyOperation({
         sessionKey: `agent:main:restart-frozen-${mode}`,
@@ -337,7 +300,7 @@ describe("embedded-agent runner run registry", () => {
     const abort = vi.fn(() => {
       throw new Error("cancel failed");
     });
-    const handle = createRunHandle({ abort });
+    const handle = createEmbeddedRunHandle({ abort });
     const operation = createReplyOperation({
       sessionKey: "agent:main:restart-throwing",
       sessionId: "session-restart-throwing",
@@ -359,7 +322,7 @@ describe("embedded-agent runner run registry", () => {
 
   it("does not bypass retained terminal ownership through compacting handle aborts", () => {
     const abort = vi.fn();
-    const handle = createRunHandle({ abort, isCompacting: true });
+    const handle = createEmbeddedRunHandle({ abort, isCompacting: true });
     const operation = createReplyOperation({
       sessionKey: "agent:main:restart-failed-compacting",
       sessionId: "session-restart-failed-compacting",
@@ -385,7 +348,7 @@ describe("embedded-agent runner run registry", () => {
   it("records active run session files in diagnostic state for heartbeat recovery", () => {
     setDiagnosticsEnabledForProcess(true);
     const sessionFile = "/tmp/openclaw-run-registry-session.jsonl";
-    const handle = createRunHandle();
+    const handle = createEmbeddedRunHandle();
 
     setActiveEmbeddedRun("session-file-diagnostics", handle, "agent:main:visible", sessionFile);
 
@@ -394,150 +357,4 @@ describe("embedded-agent runner run registry", () => {
     );
   });
 
-  it("returns a structured no-active-run queue failure", () => {
-    const outcome = queueEmbeddedAgentMessageWithOutcome("session-missing", "continue");
-
-    expect(outcome).toEqual({
-      queued: false,
-      sessionId: "session-missing",
-      reason: "no_active_run",
-      gatewayHealth: "live",
-    });
-    expect(formatEmbeddedAgentQueueFailureSummary(outcome)).toBe(
-      "queue_message_failed reason=no_active_run sessionId=session-missing gatewayHealth=live",
-    );
-  });
-
-  it("returns structured queue failures for legacy, unavailable, or compacting runs", () => {
-    const legacyQueue = vi.fn(async () => {});
-    const unavailableQueue = vi.fn(async () => {});
-    setActiveEmbeddedRun(
-      "session-not-streaming",
-      createRunHandle({ isStreaming: false, queueMessage: legacyQueue }),
-    );
-    setActiveEmbeddedRun(
-      "session-unavailable",
-      createRunHandle({
-        messageInjection: { isAvailable: () => false, queueMessage: unavailableQueue },
-      }),
-    );
-    setActiveEmbeddedRun("session-compacting", createRunHandle({ isCompacting: true }));
-
-    expect(queueEmbeddedAgentMessageWithOutcome("session-not-streaming", "continue")).toMatchObject(
-      { queued: false, reason: "not_streaming" },
-    );
-    expect(legacyQueue).not.toHaveBeenCalled();
-    expect(queueEmbeddedAgentMessageWithOutcome("session-unavailable", "continue")).toMatchObject({
-      queued: false,
-      reason: "not_streaming",
-    });
-    expect(unavailableQueue).not.toHaveBeenCalled();
-    expect(queueEmbeddedAgentMessageWithOutcome("session-compacting", "continue")).toMatchObject({
-      queued: false,
-      reason: "compacting",
-    });
-  });
-
-  it("returns runtime rejection details when async queue delivery fails", async () => {
-    setActiveEmbeddedRun("session-rejected", {
-      ...createRunHandle(),
-      queueMessage: async () => {
-        throw new Error("cannot steer a compact turn");
-      },
-    });
-
-    const outcome = await queueEmbeddedAgentMessageWithOutcomeAsync("session-rejected", "continue");
-
-    expect(outcome).toEqual({
-      queued: false,
-      sessionId: "session-rejected",
-      reason: "runtime_rejected",
-      gatewayHealth: "live",
-      errorMessage: "cannot steer a compact turn",
-    });
-    expect(formatEmbeddedAgentQueueFailureSummary(outcome)).toBe(
-      "queue_message_failed reason=runtime_rejected sessionId=session-rejected gatewayHealth=live error=cannot steer a compact turn",
-    );
-  });
-
-  it("reports accepted steering without transcript confirmation as non-replayable", async () => {
-    setActiveEmbeddedRun("session-unconfirmed", {
-      ...createRunHandle(),
-      queueMessage: async () => ({
-        transcriptCommit: "unconfirmed",
-        errorMessage: "receipt unavailable",
-      }),
-    });
-
-    const outcome = await queueEmbeddedAgentMessageWithOutcomeAsync(
-      "session-unconfirmed",
-      "continue",
-    );
-
-    expect(outcome).toEqual({
-      queued: true,
-      sessionId: "session-unconfirmed",
-      target: "embedded_run",
-      gatewayHealth: "live",
-      transcriptCommit: "unconfirmed",
-      errorMessage: "receipt unavailable",
-      enqueuedAtMs: expect.any(Number),
-    });
-  });
-
-  it("rejects transcript-commit waits for active handles without support", async () => {
-    const queueMessage = vi.fn(async () => {});
-    setActiveEmbeddedRun("session-no-transcript-wait", {
-      ...createRunHandle(),
-      queueMessage,
-    });
-
-    const outcome = await queueEmbeddedAgentMessageWithOutcomeAsync(
-      "session-no-transcript-wait",
-      "continue",
-      { waitForTranscriptCommit: true },
-    );
-
-    expect(outcome).toEqual({
-      queued: false,
-      sessionId: "session-no-transcript-wait",
-      reason: "transcript_commit_wait_unsupported",
-      gatewayHealth: "live",
-    });
-    expect(queueMessage).not.toHaveBeenCalled();
-  });
-
-  it("rejects transcript-commit waits before reply-run fallback without an active handle", async () => {
-    const queueMessage = vi.fn(async () => {});
-    const operation = createReplyOperation({
-      sessionKey: "agent:main:main",
-      sessionId: "session-reply-run",
-      resetTriggered: false,
-    });
-    operation.attachBackend({
-      kind: "embedded",
-      cancel: vi.fn(),
-      isStreaming: () => true,
-      queueMessage,
-    });
-    operation.setPhase("running");
-    const recorder = createUserTurnTranscriptRecorder({
-      input: { text: "visible group prompt", sender: { id: "user-42" } },
-      target: createTestUserTurnTranscriptTarget(),
-    });
-
-    const outcome = await queueEmbeddedAgentMessageWithOutcomeAsync(
-      "session-reply-run",
-      "completion from child",
-      { waitForTranscriptCommit: true, userTurnTranscriptRecorder: recorder },
-    );
-
-    expect(outcome).toEqual({
-      queued: false,
-      sessionId: "session-reply-run",
-      reason: "transcript_commit_wait_unsupported",
-      gatewayHealth: "live",
-    });
-    expect(queueMessage).not.toHaveBeenCalled();
-  });
 });
